[build] Include GFloppy
GFloppy has been removed from master while the development team
figures what to do with it. As gnome-utils 2.26 shipped with
gfloppy, though, we need it in the gnome-2-26 branch at the very
least.

Instead of messing the repository up more and more, let's just
bite the bullet and do what we should have done years ago: copy
gfloppy inside gnome-utils. It's not like we care *that* much
about the history -- the gfloppy repository has it.