[gsettings-tutorial] Initialize subfolders tooltip visibility
Instead of using gconf to get the value of this key, we use the
GSettings API.

It would have been easier to use g_settings_bind(), but the
BaobabRingschart object doesn't have a real property for tooltip
visibility. We can see here that implementing properties properly in
objects makes the use of GSettings easier (even though this is still
relatively easy).

A few notes:

  + We're using an accessor here: g_settings_get_boolean(). It would
    have been possible to use g_settings_get() directly:
      g_settings_get (settings, key, "b", &variable);