[gsettings-tutorial] Make the application use the gconf backend
To have a smooth migration for the code, we want to have GSettings and
gconf coexist during the port. There's a GSettings backend that uses
gconf for storage. This makes it possible to have one part of the code
using GSettings while still allowing another part of the code to use
gconf for the same key.

A few notes:

  + This backend is only available in gconf >= 2.31.1.

  + This should be removed once the port is over and migration to the
    default GSettings backend is ready.

  + GSettings won't work unless there's a schema installed for the
    settings that will get accessed.